Адаптация для git-flow

+++ master_repo.py
@@ -7,17 +7,18 @@ from icdump import base_dump, is_skipped
 
 dev_files= [] # contains relative paths
 
-for root, dirs, files in os.walk(base_dump.DEVELOPER):
+for root, dirs, files in os.walk(base_dump.DEVELOP):
     for dev_f in files:
         dev_fn= os.path.join(root, dev_f)
-        dev_fnrp= os.path.relpath(dev_fn, base_dump.DEVELOPER)
+        dev_fnrp= os.path.relpath(dev_fn, base_dump.DEVELOP)
         dev_files.append(dev_fnrp)
 
 
 for root, dirs, files in os.walk(base_dump.REPO):
+    if '.git' in root: continue # except git files
     for repo_f in files:
         repo_fn= os.path.join(root, repo_f)
-        repo_fnrp= os.path.relpath(repo_fn, base_dump.DEVELOPER) # developer file relative path
+        repo_fnrp= os.path.relpath(repo_fn, base_dump.DEVELOP) # developer file relative path
         if not repo_fnrp in dev_files:
             repo_fn= os.path.join(base_dump.REPO, repo_fnrp)
             os.remove(repo_fn)
@@ -32,12 +33,12 @@ for root, dirs, files in os.walk(base_dump.REPO):
         repo_fnrp= os.path.relpath(repo_fn, base_dump.REPO)
         repo_files.append(repo_fnrp)
 
-for root, dirs, files in os.walk(base_dump.DEVELOPER):
+for root, dirs, files in os.walk(base_dump.DEVELOP):
     for dev_f in files:
         if is_skipped(dev_f): 
             continue
         dev_fn= os.path.join(root, dev_f)
-        dev_fnrp= os.path.relpath(dev_fn, base_dump.DEVELOPER)
+        dev_fnrp= os.path.relpath(dev_fn, base_dump.DEVELOP)
         if not dev_fnrp in repo_files:
             src= os.path.join(base_dump.MASTER, dev_fnrp)
             dst= os.path.join(base_dump.REPO, dev_fnrp)
